Local

+++ app/src/main/java/com/example/universitybudgetapp/ui/components/SummaryHeader.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 
 @Composable
-fun SummaryHeader(incomeTotal: Int, expenseTotal: Int) {
+fun SummaryHeader(incomeTotal: Long, expenseTotal: Long) {
     val net = incomeTotal - expenseTotal
 
     Surface(
@@ -16,11 +16,14 @@ fun SummaryHeader(incomeTotal: Int, expenseTotal: Int) {
             .fillMaxWidth()
             .padding(bottom = 12.dp),
         color = MaterialTheme.colorScheme.surfaceVariant,
-        shape = RoundedCornerShape(12.dp), // ✅ 네 방향 동일하게 둥글게!
+        shape = RoundedCornerShape(12.dp),
         tonalElevation = 2.dp
     ) {
         Column(modifier = Modifier.padding(horizontal = 20.dp, vertical = 16.dp)) {
-            Row(modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween) {
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween
+            ) {
                 Text("총 수입", style = MaterialTheme.typography.bodyMedium)
                 Text(
                     text = "+%,d원".format(incomeTotal),
@@ -31,7 +34,10 @@ fun SummaryHeader(incomeTotal: Int, expenseTotal: Int) {
 
             Spacer(modifier = Modifier.height(8.dp))
 
-            Row(modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween) {
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween
+            ) {
                 Text("총 지출", style = MaterialTheme.typography.bodyMedium)
                 Text(
                     text = "-%,d원".format(expenseTotal),
@@ -40,11 +46,14 @@ fun SummaryHeader(incomeTotal: Int, expenseTotal: Int) {
                 )
             }
 
-            Spacer(modifier = Modifier.height(12.dp)) // ✅ Divider 간격 여유롭게
+            Spacer(modifier = Modifier.height(12.dp))
             Divider()
             Spacer(modifier = Modifier.height(12.dp))
 
-            Row(modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween) {
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween
+            ) {
                 Text("합계", style = MaterialTheme.typography.bodyMedium)
                 Text(
                     text = "%,d원".format(net),
